+++ fe/fe-core/src/test/java/org/apache/doris/nereids/memo/MemoRewriteTest.java
@@ -57,57 +57,6 @@ class MemoRewriteTest implements PatternMatchSupported {
 
     private final ConnectContext connectContext = MemoTestUtils.createConnectContext();
 
-    private final LogicalJoin<LogicalOlapScan, LogicalOlapScan> logicalJoinAB = new LogicalJoin<>(JoinType.INNER_JOIN,
-            PlanConstructor.newLogicalOlapScan(0, "A", 0),
-            PlanConstructor.newLogicalOlapScan(1, "B", 0));
-
-    private final LogicalJoin<LogicalJoin<LogicalOlapScan, LogicalOlapScan>, LogicalOlapScan> logicalJoinABC = new LogicalJoin<>(
-            JoinType.INNER_JOIN, logicalJoinAB, PlanConstructor.newLogicalOlapScan(2, "C", 0));
-
-    /**
-     * Original:
-     * Group 0: LogicalOlapScan C
-     * Group 1: LogicalOlapScan B
-     * Group 2: LogicalOlapScan A
-     * Group 3: Join(Group 1, Group 2)
-     * Group 4: Join(Group 0, Group 3)
-     * <p>
-     * Then:
-     * Copy In Join(Group 2, Group 1) into Group 3
-     * <p>
-     * Expected:
-     * Group 0: LogicalOlapScan C
-     * Group 1: LogicalOlapScan B
-     * Group 2: LogicalOlapScan A
-     * Group 3: Join(Group 1, Group 2), Join(Group 2, Group 1)
-     * Group 4: Join(Group 0, Group 3)
-     */
-    @Test
-    public void testInsertSameGroup() {
-        PlanChecker.from(MemoTestUtils.createConnectContext(), logicalJoinABC)
-                .transform(
-                        // swap join's children
-                        logicalJoin(logicalOlapScan(), logicalOlapScan()).then(joinBA ->
-                                new LogicalProject<>(Lists.newArrayList(joinBA.getOutput()),
-                                        new LogicalJoin<>(JoinType.INNER_JOIN, joinBA.right(), joinBA.left()))
-                        ))
-                .checkGroupNum(6)
-                .checkGroupExpressionNum(7)
-                .checkMemo(memo -> {
-                    Group root = memo.getRoot();
-                    Assertions.assertEquals(1, root.getLogicalExpressions().size());
-                    GroupExpression joinABC = root.getLogicalExpression();
-                    Assertions.assertEquals(2, joinABC.child(0).getLogicalExpressions().size());
-                    Assertions.assertEquals(1, joinABC.child(1).getLogicalExpressions().size());
-                    GroupExpression joinAB = joinABC.child(0).getLogicalExpressions().get(0);
-                    GroupExpression project = joinABC.child(0).getLogicalExpressions().get(1);
-                    GroupExpression joinBA = project.child(0).getLogicalExpression();
-                    Assertions.assertTrue(joinAB.getPlan() instanceof LogicalJoin);
-                    Assertions.assertTrue(joinBA.getPlan() instanceof LogicalJoin);
-                });
-
-    }
-
     @Test
     public void initByOneLevelPlan() {
         OlapTable table = PlanConstructor.newOlapTable(0, "a", 1);
